test/requirements.txt
python-hcl2>=0.3.0
pyyaml>=6.0
boto3>=1.26.0 